requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
pydantic>=2.9.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
openai==1.30.1
faker==25.2.0
python-dotenv==1.0.1
